added chat timestamps module cuz its cool ig

+++ src/main/java/me/bakje/bakjedev/bakjedev/module/Settings/NumberSetting.java
@@ -21,7 +21,7 @@ public class NumberSetting extends Setting{
         return value;
     }
 
-    public double getValueFloat() {
+    public float getValueFloat() {
         return (float)value;
     }
